requirements.txt
abydos==0.5.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
language-tool-python==2.8
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.26.4
packaging==24.0
python-dotenv==1.0.1
regex==2024.5.15
requests==2.32.2
textblob==0.18.0.post0
tqdm==4.66.4
urllib3==2.2.1
Werkzeug==3.0.3
